fix: move exchanges data to constants
