+++ constants/exchanges.ts
@@ -1,0 +1,20 @@
+const exchanges = [
+  {
+    name: 'Uniswap',
+    url:
+      'https://v2.info.uniswap.org/pair/0xbaeca7c35346a8d31811ef971f38603012a12c1e',
+    logo: '/svgs/uniswap.svg',
+  },
+  {
+    name: 'Coinex',
+    url: 'https://www.coinex.com/exchange/oxen-usdt',
+    logo: '/svgs/coinex.svg',
+  },
+  {
+    name: 'Tradeogre',
+    url: 'https://tradeogre.com/exchange/OXEN-BTC',
+    logo: '/svgs/tradeogre.svg',
+  },
+];
+
+export default exchanges;
