+++ constants/index.ts
@@ -4,5 +4,6 @@ import NAVIGATION from '@/constants/navigation';
 import SEARCH from '@/constants/search';
 import TOS from './tos';
 import UI from '@/constants/ui';
+import EXCHANGES from '@/constants/exchanges';
 
-export { CMS, METADATA, NAVIGATION, SEARCH, TOS, UI };
+export { CMS, METADATA, NAVIGATION, SEARCH, TOS, UI, EXCHANGES };
